voiceapi/requirements.txt
alabaster==0.7.16
antlr4-python3-runtime==4.8
asgiref==3.7.2
Babel==2.14.0
bcrypt==4.1.2
bitarray==2.9.2
blinker==1.7.0
boto3==1.34.2
botocore==1.34.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Cython==3.0.8
defusedxml==0.8.0rc2
Django==4.2.6
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docutils==0.20.1
fairseq==0.12.2
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.2.0
gevent==23.9.1
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-texttospeech==2.16.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
h11==0.14.0
hydra-core==1.0.7
idna==3.4
imagesize==1.4.1
indic-nlp-library==0.92
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
mock==5.1.0
Morfessor==2.0.6
mpmath==1.3.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.0.6
packaging==23.2
pandas==2.2.0
paramiko==3.4.0
phonenumbers==8.13.29
portalocker==2.8.2
progressbar2==4.3.2
proto-plus==1.23.0
protobuf==4.21.12
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydload==1.0.9
pydub==0.25.1
Pygments==2.17.2
PyJWT==1.7.1
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-utils==3.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.9.0
sacrebleu==2.4.0
sacremoses==0.1.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
Sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urduhack==0.1.4
urllib3==2.0.7
uvicorn==0.24.0.post1
websockets==12.0
Werkzeug==3.0.1
django-redis==5.4.0
zope.event==5.0
zope.interface==6.1
annotated-types==0.6.0
asgiref==3.7.2
astroid==2.5
attrs==23.2.0
boto3==1.34.60
botocore==1.34.60
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==7.1.2
clickclick==20.10.2
connexion==2.9.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Flask==1.1.4
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.1
idna==3.6
inflection==0.5.1
isodate==0.6.1
isort==5.13.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
jsonschema==3.2.0
jwcrypto==1.5.4
lazy-object-proxy==1.10.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
pandas==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==1.6.1
pydantic_core==2.16.3
pydub==0.25.1
PyJWT==2.8.0
pylint==2.6.0
pyrsistent==0.16.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
s3transfer==0.10.0
scipy==1.12.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soundfile==0.12.1
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==1.0.1
wrapt==1.12.1
asgiref==3.7.2
audioread==3.0.1
beautifulsoup4==4.12.3
boto3==1.34.35
botocore==1.34.35
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
google==3.0.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-texttospeech==2.16.1
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
gunicorn==23.0.0
h11==0.14.0
idna==3.6
imageio-ffmpeg==0.6.0
jmespath==1.0.1
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
msgpack==1.0.8
mysqlclient==2.2.3
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
platformdirs==4.2.0
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
pydub==0.25.1
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
sqlparse==0.4.4
threadpoolctl==3.4.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.27.0.post1
